voice_agent/requirements.txt
groq
edge-tts
SpeechRecognition
pyaudio
setuptools
pyautogui
AppOpener
python-dotenv
sounddevice
numpy
scipy
colorama
pygame
customtkinter
packaging
WMI
Pillow
psutil
requests
beautifulsoup4
pyperclip
win10toast
schedule
